fix(BigList): refresh vis map on slice changes to maintain focus
Temporarily resolves focus desync after slice changes, but this exposes
a larger design issue explained in the comments. Ultimately, a refactor
will need to be done to ensure removing and appending can be done
safely and to ensure that maintaining a fresh visual map is done without
redundancy. (possible something along the lines of scheduling visual map
refreshes with focus syncs on every tick after a tree manip method was
called).

This also exposes the suble annoyance of removing all children since the
public `children` is a getter and thus, iterating over it is live.
Should implement a `replaceChildren` method like the browser.