fix(Renderer): resizes clear all display after cursor
If the width of a bordered box was stdout cols, after a resize to make
the screen more narrow, the terminal treats the old width as a single
line and wraps it, but when clearing from the cursor to the end of that
row it does not clear the wrapped portion. Therefore, we end up with
stale lines below the actual canvas.  This clears everything from the
cursor to the end of the screen on resizes to clip this away.