fix-bug(Scheduler): fast animations/transitions blocking listener cbs
The Scheduler does not allow event listener callbacks to execute when
composing the layout to be painted onto the screen.  This was to prevent
the pollution of the state of the app during rendering in case a key was
pressed during this process.  This was mostly important if there was
significant lag due to a massive dom tree, but could also theoretically
pollute the DOM state during a fast calculation.  Therefore, the callbacks
are batched and executed when possible.  The way it was written only
allowed the cbs to run after the updates had ran, and in the case of a
animation/transition below the debounceMs, the cbs were always blocked
from running.  This changes the sequence to allow execution of event
callbacks during any animation/transition.