feat(text): make TextElement width sized to its given textContent
This fundamentally changes text element behavior as well.  Before, text
width behaved more like the browser, taking up 100% width of its
container, and it now only takes up the width of its own width.

This method is more practical for a TUI in my opinion and it also needed
to be done because Yoga's measureFunc supplies a width that is sometimes
too wide or too short, which means that expecting background colors to
span the entire width is unpredictable.  This also allows text to be
aligned on the horizontal axis through the parent element, which is more
ergonomic than forcing the textStyle.align to control horizontal axis
position.